Combine all Enum classes/modules into one `enumerations` module.
